requirements.txt
PySimpleGUI
getpass
Pillow
stability-sdk
